+++ lib/utils/index.ts
@@ -1,0 +1,2 @@
+export { getGradeColor, getGradeBgColor, getGradeClasses } from './grades';
+export { formatRelativeDate, formatShortDate } from './dates';